Separate sendTableMessage in utils/
